move version.lua to kissmp directory to prevent filename conflicts with 'version'
